+++ api/models/post.py
@@ -7,8 +7,6 @@ class Post(Document):
     author = ReferenceField(User)
     title = StringField(required=True)
     content = StringField(required=True)
-    updated = BooleanField()
     dateTime = DateTimeField()
-    updateTime = DateTimeField()
     like_counter = IntField(min_value=0)
     dislike_counter = IntField(min_value=0)
